requirements.txt
requests
requests[socks]
bs4
user_agent
youtube_dl
requests_toolbelt
telethon
tenacity
pyrogram
tgcrypto
aiohttp
yarl
pyrogram
Pysocks
aiohttp_socks
mega
mega.py